Added the Android Library for the project.
Create an Android Library containing an Activity that will display a joke passed to it as an intent extra. Wire up project dependencies so that the button can now pass the joke from the Java Library to the Android Library.

I have added the cardview dependency for the andorid library project. Added few extra jokes and i will add more jokes over the next commits.